pyproject.toml
[project]
name = "data-warehouse-migrate"
version = "0.1.0"
description = "数仓迁移工具，支持从阿里云MaxCompute迁移到Google Cloud BigQuery"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-bigquery>=3.11.0",
    "pyodps>=0.11.0",
    "click>=8.1.0",
    "pandas>=2.0.0",
    "pyarrow>=12.0.0",
    "tqdm>=4.65.0",
    "python-dotenv>=1.0.0",
    "pandas-gbq>=0.26.1",
]

[project.scripts]
data-warehouse-migrate = "data_warehouse_migrate.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
